requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cv2-tools==2.4.0
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
image-classifiers==1.0.0
ImageHash==4.2.0
importlib-metadata==1.7.0
joblib==0.16.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lmdb==1.2.1
Markdown==3.2.2
matplotlib==3.3.0
msgpack==1.0.2
msgpack-numpy==0.4.7.1
networkx==2.5
nltk==3.6.1
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.0.5
Pillow==7.2.0
protobuf==3.12.2
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pygraphviz==1.6
pyparsing==2.4.7
python-constraint==1.4.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.11.1
six==1.14.0
sklearn==0.0
tabulate==0.8.9
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorpack==0.11
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.60.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
